Completed remaining functions (85%).  Debugged code over the last 8-10 hours. Styled cards.
Over the last 8 - 10 hours I have completed the remaining functions that I created placeholders for in the previous commit.  Most if not all of them did not function correctly at first.  De-bugged the code for each function via console logs and comments.  Figured out how to iterate through the object data returned from the third party AI.  Using Next5Days to loop through the forecastData that returns in a list of objects.

Searched through third party API docs and located the list of data returned and how to pull it.  Used this in my current and next5days forecast to implement the icon, date, temp, wind and humidity.

Still need to grab the city name to label the current day.

May switch to city, state fetch call if all else is working in the future.

Moved buttons to create in nav bar since the search field is there as well.

Applied styles in custom css to get the cards to display horizontally.