Pass City name through fetch as arg for current day.  Located city name bug in local storage. Updated html <h1>
I am now passing the city name from the fetch call through as an argument for current day function to display the name above.  Located city name bug in local storage, every time a button event listener is called the list appends the city name.  This will continuously grow.  Currently not affecting anything adversely but would like to fix it.  Updated the Nav bar to remove useless <a href's that were in the top left.  Made it a single <h1> title.